feat: rollback action to the module
